requirements/tests.txt
flake8
coverage
django-discover-runner
mock